+++ Client/src/Miscellaneous/Vector.cpp
@@ -3,7 +3,7 @@
 
 // Methods
 float Vector2::DistanceToFast(Vector2 other) const { return abs(x - other.x) + abs(y - other.y); }
-float Vector2::DistanceTo(Vector2 other) const { return sqrt(pow(other.x - x, 2) + pow(other.y - y, 2)); }
+float Vector2::DistanceTo(Vector2 other) const { return (float)sqrt(pow(other.x - x, 2) + pow(other.y - y, 2)); }
 float Vector2::Dot(Vector2 other) const { return other.x * x + other.y * y; }
 float Vector2::Magnitude() const { return sqrt(x * x + y * y); }
 Vector2 Vector2::Normalize() { float mag = Magnitude(); if (mag == 0) return Vector2(0.0f, 0.0f); return Vector2(x / mag, y / mag); }
